refactor(split): move central sensor handling out of Bluetooth directory
